pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelspointspolygons"
version = "0.1.0"
description = "The P3 Dataset: Pixels, Points and Polygons for Multimodal Building Vectorization"
authors = [
    { name = "Raphael Sulzer", email = "raphael.sulzer.1@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "albumentations",
    "colorlog",
    "descartes==1.1.0",
    "huggingface_hub",
    "hydra-core",
    "laspy[lazrs]",
    "matplotlib",
    # "numpy==1.25.2", # this line is only necessary when installing with open3d, i.e. pytorch 2.2, not necessary with the newest pytorch
    #"open3d==0.19.0",
    "pandas",
    "pycocotools",
    "rasterio",
    "scikit-image",
    "scikit-learn",
    "shapely",
    "tensorboard",
    "timm",
    "torchmetrics",
    "transformers",
    "wandb",
    # only for hisup
    #"Cython",
    # only for FFL preprocessing
    #"kornia",
    #"skan",
    #"jsmin",
    # only for FFL ASM polygonization (default is ACM)
    # "torch-scatter==2.1.2" ## takes forever so making it optional
]


[tool.setuptools]
packages = ["pixelspointspolygons"]

[tool.setuptools.package-data]
pixelspointspolygons = ["conf/*.yaml"]
